refactor login to centralize all login features in one service
